backend/requirements.txt
asgiref==3.4.1
colorama==0.4.4
coverage==6.0.1
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.12.4
green==3.3.0
gunicorn==20.1.0
lxml==4.6.3
psycopg2==2.9.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
Unidecode==1.3.2
whitenoise==5.3.0
